requirements.txt
Flask==2.3.2
   tensorflow==2.16.1
   Pillow==10.0.0
   numpy==1.25.2